pyproject.toml
[tool.poetry]
name = "ntpan"
version = "0.1.0"
description = ""
authors = ["checktheroads <matt@allroads.io>"]

[tool.poetry.dependencies]
python = "^3.7"
scrapli = {extras = ["asyncssh"], version = "^2020.7.26"}
httpx = "^0.14.0"
xmltodict = "^0.12.0"
pydantic = {extras = ["email"], version = "^1.6.1"}
emails = "^0.6"
apscheduler = "^3.6.3"
Jinja2 = "^2.11.2"
loguru = "^0.5.1"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
mypy = "^0.782"
isort = "^5.3.2"

[tool.black]
line-length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

